@@ -12,15 +12,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.3",
-    "@angular/common": "^4.3.1",
-    "@angular/compiler": "^4.3.1",
-    "@angular/core": "^4.3.1",
-    "@angular/forms": "^4.3.1",
-    "@angular/http": "^4.3.1",
-    "@angular/platform-browser": "^4.3.1",
-    "@angular/platform-browser-dynamic": "^4.3.1",
-    "@angular/router": "^4.3.1",
+    "@angular/animations": "^5.2.3",
+    "@angular/common": "^5.2.3",
+    "@angular/compiler": "^5.2.3",
+    "@angular/core": "^5.2.3",
+    "@angular/forms": "^5.2.3",
+    "@angular/http": "^5.2.3",
+    "@angular/platform-browser": "^5.2.3",
+    "@angular/platform-browser-dynamic": "^5.2.3",
+    "@angular/platform-server": "^5.2.3",
+    "@angular/router": "^5.2.3",
     "@types/dropzone": "^5.0.1",
     "angular2-text-mask": "^8.0.4",
     "animate.css": "^3.5.2",
@@ -33,21 +34,21 @@
     "jquery-sparkline": "^2.4.0",
     "jqvmap": "^1.5.1",
     "material-design-iconic-font": "^2.2.0",
-    "ng-autosize": "^1.0.1",
+    "ng-autosize": "^1.1.0",
     "ng2-nouislider": "^1.7.0",
     "ng2-select2": "^1.0.0-beta.10",
     "ng2-table": "^1.3.2",
     "ngx-bootstrap": "^2.0.0-beta.6",
-    "ngx-perfect-scrollbar": "^4.8.4",
+    "ngx-perfect-scrollbar": "^5.0.0",
     "nouislider": "^10.1.0",
     "popper.js": "^1.13.0",
-    "rxjs": "^5.4.2",
+    "rxjs": "^5.5.6",
     "select2": "^4.0.4",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular/cli": "^1.6.7",
-    "@angular/compiler-cli": "^4.4.4",
+    "@angular/compiler-cli": "^5.2.3",
     "@angular/language-service": "^4.3.1",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
@@ -66,6 +67,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "2.3.4"
+    "typescript": "~2.6.2"
   }
 }
